requirements.txt
# Core Framework
fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.10.4
pydantic-settings==2.7.0

# Authentication & Security
python-jose[cryptography]==3.4.0
bcrypt==5.0.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.21

# Database
pymongo==4.10.1
dnspython==2.7.0

# Utilities
python-dotenv==1.0.1
email-validator==2.2.0
pytz==2025.2

# AI/LLM Integration
google-genai==1.2.0
tenacity==9.0.0

# AWS SDK
boto3==1.35.81
botocore==1.35.81

# Additional AWS/Production packages
gunicorn==23.0.0
mangum==0.20.0

# Email Service
argon2-cffi==25.1.0